update thoughts after creating new from api
