+++ code/src/App.js
@@ -1,13 +1,22 @@
-import React from 'react'
+import React, { useEffect, useState }  from 'react'
 
-import ThoughtForm from 'components/ThoughtForm'
-import ThoughtList from 'components/ThoughtList'
+import CreateThought from 'components/CreateThought'
+import ShowThoughtList from 'components/ShowThoughtList'
 
 export const App = () => {
+
+  const [thoughtList, setThoughtList] = useState([])
+
+  useEffect(() => {
+    fetch("https://happy-thoughts-technigo.herokuapp.com/thoughts")
+      .then(res => res.json())
+      .then(data => setThoughtList(data))
+  }, [])
+
   return (
     <div>
-      <ThoughtForm />
-      <ThoughtList />
+      <CreateThought setThoughts={setThoughtList} />
+      <ShowThoughtList thoughtList={thoughtList}/>
     </div>
   )
 }
